requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=1.4.0=pypi_0
anyio=3.7.1=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.4.0=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.1.0=pypi_0
babel=2.12.1=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
blas=2.116=mkl
blas-devel=3.9.0=16_linux64_mkl
bleach=6.0.0=pypi_0
boost=1.70.0=py38h9de70de_1
boost-cpp=1.70.0=ha2d47e9_1
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7f98852_4
c-ares=1.19.1=hd590300_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.0=pypi_0
cairo=1.16.0=h18b612c_1001
certifi=2023.5.7=py38h06a4308_0
cffi=1.15.1=py38h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
comm=0.1.4=pypi_0
contourpy=1.0.7=pypi_0
cryptography=39.0.1=py38h9ce1e76_0
cudatoolkit=11.3.1=h9edb442_11
curl=7.87.0=h5eee18b_0
cycler=0.11.0=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
easydict=1.10=pypi_0
exceptiongroup=1.1.3=pypi_0
executing=1.2.0=pypi_0
expat=2.5.0=hcb278e6_1
fastjsonschema=2.18.0=pypi_0
fontconfig=2.14.1=hef1e5e3_0
fonttools=4.39.4=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.10.4=h0708190_1
giflib=5.2.1=h0b41bf4_3
glew=2.1.0=h9c3ff4c_2
glib=2.69.1=h4ff587b_1
glm=0.9.9.8=h00ab1b0_0
google-auth=2.18.1=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.54.2=pypi_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
hdf4=4.2.15=h10796ff_3
hdf5=1.12.1=nompi_h2750804_100
icu=58.2=hf484d3e_1000
idna=3.4=py38h06a4308_0
importlib-metadata=6.6.0=pypi_0
importlib-resources=5.12.0=pypi_0
ipykernel=6.25.2=pypi_0
ipython=8.12.2=pypi_0
ipython-genutils=0.2.0=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.0=pypi_0
jinja2=3.1.2=py38h06a4308_0
joblib=1.1.1=py38h06a4308_0
jpeg=9e=h0b41bf4_3
json5=0.9.14=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.19.0=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
jupyter-client=8.3.1=pypi_0
jupyter-core=5.3.1=pypi_0
jupyter-events=0.7.0=pypi_0
jupyter-lsp=2.2.0=pypi_0
jupyter-server=1.24.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab=3.0.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.25.0=pypi_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=pypi_0
krb5=1.19.3=h3790be6_0
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=16_linux64_mkl
libcblas=3.9.0=16_linux64_mkl
libclang=10.0.1=default_hb85057a_2
libcurl=7.87.0=h91b91d3_0
libdeflate=1.17=h0b41bf4_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.12=h8f2d780_0
libexpat=2.5.0=hcb278e6_1
libffi=3.3=he6710b0_2
libgcc-ng=12.2.0=h65d4601_19
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libglu=9.0.0=hac7e632_1002
liblapack=3.9.0=16_linux64_mkl
liblapacke=3.9.0=16_linux64_mkl
libllvm10=10.0.1=he513fc3_3
libnetcdf=4.8.1=nompi_hb3fd0d9_101
libnghttp2=1.46.0=hce63b2e_0
libopenblas=0.3.21=pthreads_h78a6416_3
libpng=1.6.39=h5eee18b_0
libpq=12.9=h16c4e8d_3
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=12.2.0=h46fd767_19
libtiff=4.5.0=h6a678d5_2
libuuid=2.38.1=h0b41bf4_0
libuv=1.44.2=h166bdaf_0
libwebp=1.2.4=h1daa5a0_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h0b41bf4_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
libzip=1.8.0=h5cef20c_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=hcb278e6_0
markdown=3.4.3=pypi_0
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=3.0.1=pypi_0
mkl=2022.1.0=h84fe81f_915
mkl-devel=2022.1.0=ha770c72_916
mkl-include=2022.1.0=h84fe81f_915
nbclassic=0.5.6=pypi_0
nbclient=0.8.0=pypi_0
nbconvert=7.8.0=pypi_0
nbformat=5.9.2=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pypi_0
networkx=2.8.4=py38h06a4308_1
notebook-shim=0.2.3=pypi_0
nspr=4.35=h27087fc_0
nss=3.74=h0370c37_0
numpy=1.23.0=pypi_0
oauthlib=3.2.2=pypi_0
openbabel=3.1.1=py38hd2c4bc0_2
openssl=1.1.1t=h7f8727e_0
overrides=7.4.0=pypi_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.1=py38h01efb38_1
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=py38h6a678d5_0
pip=23.0.1=py38h06a4308_0
pixman=0.38.0=h516909a_1003
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.10.0=pypi_0
ply=3.11=py_1
pmw=2.0.1=py38h578d9bd_1007
pooch=1.4.0=pyhd3eb1b0_0
prometheus-client=0.17.1=pypi_0
prompt-toolkit=3.0.39=pypi_0
protobuf=4.23.1=pypi_0
psutil=5.9.5=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycairo=1.23.0=py38h190342e_0
pycparser=2.21=pyhd3eb1b0_0
pyg=2.0.4=py38_torch_1.10.0_cu113
pygments=2.16.1=pypi_0
pymol-open-source=2.5.0=py38h9174eab_6
pyopenssl=23.0.0=py38h06a4308_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.7=py38h6a678d5_1
pyqt5-sip=12.11.0=py38h6a678d5_1
pysocks=1.7.1=py38h06a4308_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-json-logger=2.0.7=pypi_0
python-lmdb=1.4.1=py38h8dc9893_0
python-louvain=0.15=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.10.1=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-cluster=1.6.0=py38_torch_1.10.0_cu113
pytorch-mutex=1.0=cuda
pytorch-scatter=2.0.9=py38_torch_1.10.0_cu113
pytorch-sparse=0.6.13=py38_torch_1.10.0_cu113
pytorch-spline-conv=1.2.1=py38_torch_1.10.0_cu113
pytz=2023.3=pyhd8ed1ab_0
pyyaml=6.0=py38h5eee18b_1
pyzmq=25.1.1=pypi_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
rdkit=2019.09.3=py38hb31dc5d_0
readline=8.2=h5eee18b_0
referencing=0.30.2=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.10.3=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.2.2=py38h6a678d5_0
scipy=1.10.1=py38h59b608b_3
seaborn=0.13.0=pypi_0
send2trash=1.8.2=pypi_0
setuptools=58.0.4=py38h06a4308_0
sip=6.6.2=py38hfa26641_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.2=pypi_0
tbb=2021.7.0=h924138e_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=pypi_0
tornado=6.3.3=pypi_0
tqdm=4.65.0=py38hb070fc8_0
traitlets=5.10.0=pypi_0
typing_extensions=4.5.0=pyha770c72_0
uri-template=1.3.0=pypi_0
urllib3=1.26.15=py38h06a4308_0
wcwidth=0.2.6=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.6.3=pypi_0
werkzeug=2.3.4=pypi_0
wheel=0.38.4=py38h06a4308_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.8.4=h8ee46fc_1
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.10=h5eee18b_1
yacs=0.1.6=pyhd3eb1b0_1
yaml=0.2.5=h7b6447c_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
